requirements.txt
Flask
Flask-WTF
python-dotenv
Flask-SQLAlchemy
gunicorn
bleach
